+++ djangosaml2_v0181post/templatetags/idplist.py
@@ -14,8 +14,8 @@
 
 from django import template
 
-from djangosaml2.conf import config_settings_loader
-from djangosaml2.utils import available_idps
+from djangosaml2_v0181post.conf import config_settings_loader
+from djangosaml2_v0181post.utils import available_idps
 
 register = template.Library()
